requirements.txt
supabase>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
playwright>=1.40.0
